vertexai-embeddings-retrieval-langchain/requirements.txt
langchain
langchain-community
langchain-google-vertexai
pypdf
python-dotenv
redis
redisvl
urllib3==1.26.18
tiktoken